backend/app/services/production/service_integration.py
"""
Production Service Integration - ZERO MOCK DATA
Updates existing services to use production implementations
"""

import os
import logging
from typing import Dict, List, Optional, Any

logger = logging.getLogger(__name__)

def update_existing_services():
    """Update existing services to use production implementations - NO MOCK DATA"""
    
    # This function documents the changes needed to integrate production services
    integration_plan = {
        'services_to_update': [
            'backend/app/services/google_forms_service.py',
            'backend/app/services/microsoft_graph_service.py', 
            'backend/app/services/ai_analysis_service.py',
            'backend/app/services/template_converter.py',
            'backend/app/services/automated_report_system.py'
        ],
        'changes_summary': {
            'google_forms_service.py': 'Replace with production/google_forms_service.py - ZERO mock data',
            'microsoft_graph_service.py': 'Replace with production/microsoft_graph_service.py - Real API integration',
            'ai_analysis_service.py': 'Replace with production/ai_analysis_service.py - Real OpenAI GPT-4',
            'template_converter.py': 'Replace with production/template_converter_service.py - Real document processing',
            'automated_report_system.py': 'Replace with production/automated_report_system.py - Complete integration'
        },
        'environment_variables_required': [
            'GOOGLE_CLIENT_ID',
            'GOOGLE_CLIENT_SECRET', 
            'GOOGLE_REDIRECT_URI',
            'MICROSOFT_CLIENT_ID',
            'MICROSOFT_CLIENT_SECRET',
            'MICROSOFT_TENANT_ID',
            'MICROSOFT_REDIRECT_URI',
            'OPENAI_API_KEY',
            'TEMPLATES_DIR',
            'OUTPUT_DIR'
        ]
    }
    
    return integration_plan

# Production Environment Configuration
PRODUCTION_ENV_CONFIG = """
# Production Environment Variables - ZERO MOCK DATA

# Google Forms API Configuration
GOOGLE_CLIENT_ID=your_google_client_id_here
GOOGLE_CLIENT_SECRET=your_google_client_secret_here
GOOGLE_REDIRECT_URI=http://localhost:3000/auth/google/callback
GOOGLE_TOKENS_DIR=./tokens/google/

# Microsoft Graph API Configuration  
MICROSOFT_CLIENT_ID=your_microsoft_client_id_here
MICROSOFT_CLIENT_SECRET=your_microsoft_client_secret_here
MICROSOFT_TENANT_ID=your_microsoft_tenant_id_here
MICROSOFT_REDIRECT_URI=http://localhost:3000/auth/microsoft/callback
MICROSOFT_TOKENS_DIR=./tokens/microsoft/

# OpenAI Configuration
OPENAI_API_KEY=your_openai_api_key_here
OPENAI_API_BASE=https://api.openai.com/v1
OPENAI_MODEL=gpt-4
OPENAI_MAX_TOKENS=2000
OPENAI_TEMPERATURE=0.3

# Template and Output Configuration
TEMPLATES_DIR=./templates/
OUTPUT_DIR=./generated_reports/
DEFAULT_TEMPLATE_PATH=./templates/Temp1.docx

# Report Generation Settings
MAX_CONCURRENT_REPORTS=5

# Database Configuration (Production)
DATABASE_URL=postgresql://username:password@localhost:5432/production_db
REDIS_URL=redis://localhost:6379/0

# Security Configuration
SECRET_KEY=your_super_secure_secret_key_here
JWT_SECRET_KEY=your_jwt_secret_key_here
JWT_ACCESS_TOKEN_EXPIRES=3600

# Logging Configuration
LOG_LEVEL=INFO
LOG_FILE=./logs/production.log
"""

# Production Import Updates
PRODUCTION_IMPORTS = {
    'google_forms_service': 'from app.services.production.google_forms_service import GoogleFormsService',
    'microsoft_graph_service': 'from app.services.production.microsoft_graph_service import MicrosoftGraphService', 
    'ai_analysis_service': 'from app.services.production.ai_analysis_service import AIAnalysisService',
    'template_converter_service': 'from app.services.production.template_converter_service import TemplateConverterService',
    'automated_report_system': 'from app.services.production.automated_report_system import ProductionAutomatedReportSystem'
}

# Production Dependencies
PRODUCTION_DEPENDENCIES = """
# Additional Production Dependencies
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
msal==1.24.1
aiofiles==23.2.1
httpx==0.25.2
openai==1.3.5
python-docx==0.8.11
"""

def generate_production_requirements():
    """Generate production requirements.txt - NO MOCK DATA"""
    base_requirements = [
        "Flask==2.3.3",
        "Flask-SQLAlchemy==3.0.5", 
        "Flask-JWT-Extended==4.5.3",
        "Flask-Migrate==4.0.5",
        "Flask-CORS==4.0.0",
        "psycopg2-binary==2.9.7",
        "redis==5.0.1",
        "celery==5.3.4",
        "python-dotenv==1.0.0",
        "requests==2.31.0",
        "pandas==2.1.3",
        "openpyxl==3.1.2",
        "Pillow==10.1.0",
        "Werkzeug==2.3.7"
    ]
    
    production_requirements = [
        "google-auth==2.23.4",
        "google-auth-oauthlib==1.1.0", 
        "google-auth-httplib2==0.1.1",
        "google-api-python-client==2.108.0",
        "msal==1.24.1",
        "aiofiles==23.2.1",
        "httpx==0.25.2",
        "openai==1.3.5",
        "python-docx==0.8.11",
        "lxml==4.9.3"
    ]
    
    all_requirements = base_requirements + production_requirements
    
    return "\n".join(all_requirements)

# Production Service Registry
class ProductionServiceRegistry:
    """Registry for all production services - ZERO MOCK DATA"""
    
    def __init__(self):
        self.services = {}
        self.initialized = False
    
    def register_services(self):
        """Register all production services - NO MOCK DATA"""
        try:
            from app.services.production.google_forms_service import GoogleFormsService
            from app.services.production.microsoft_graph_service import MicrosoftGraphService
            from app.services.production.ai_analysis_service import AIAnalysisService
            from app.services.production.template_converter_service import TemplateConverterService
            from app.services.production.automated_report_system import ProductionAutomatedReportSystem
            
            self.services['google_forms'] = GoogleFormsService()
            self.services['microsoft_graph'] = MicrosoftGraphService() 
            self.services['ai_analysis'] = AIAnalysisService()
            self.services['template_converter'] = TemplateConverterService()
            self.services['automated_reports'] = ProductionAutomatedReportSystem()
            
            self.initialized = True
            logger.info("All production services registered successfully - ZERO mock data")
            
        except Exception as e:
            logger.error(f"Error registering production services: {str(e)}")
            raise Exception(f"Service registration failed: {str(e)}")
    
    def get_service(self, service_name: str):
        """Get a registered production service - NO MOCK DATA"""
        if not self.initialized:
            self.register_services()
        
        if service_name not in self.services:
            raise ValueError(f"Service '{service_name}' not found in registry")
        
        return self.services[service_name]
    
    def get_all_services(self) -> Dict[str, Any]:
        """Get all registered services - NO MOCK DATA"""
        if not self.initialized:
            self.register_services()
        
        return self.services.copy()

# Global service registry instance
production_registry = ProductionServiceRegistry()

def get_production_service(service_name: str):
    """Get a production service by name - NO MOCK DATA"""
    return production_registry.get_service(service_name)

def initialize_production_services():
    """Initialize all production services - NO MOCK DATA"""
    production_registry.register_services()
    logger.info("Production services initialization complete - ZERO mock data")

# Production Verification
def verify_production_readiness() -> Dict[str, Any]:
    """Verify that all production services are ready - NO MOCK DATA"""
    verification_results = {
        'status': 'checking',
        'services_ready': {},
        'environment_variables': {},
        'dependencies': {},
        'overall_ready': False
    }
    
    # Check environment variables
    required_env_vars = [
        'GOOGLE_CLIENT_ID', 'GOOGLE_CLIENT_SECRET',
        'MICROSOFT_CLIENT_ID', 'MICROSOFT_CLIENT_SECRET', 'MICROSOFT_TENANT_ID',
        'OPENAI_API_KEY'
    ]
    
    for var in required_env_vars:
        verification_results['environment_variables'][var] = bool(os.getenv(var))
    
    # Check service initialization
    try:
        initialize_production_services()
        services = production_registry.get_all_services()
        
        for service_name, service in services.items():
            verification_results['services_ready'][service_name] = {
                'initialized': True,
                'mock_mode': getattr(service, 'mock_mode', None)
            }
    except Exception as e:
        verification_results['services_ready']['error'] = str(e)
    
    # Check dependencies
    try:
        import google.auth
        import msal
        import openai
        import httpx
        verification_results['dependencies']['google_auth'] = True
        verification_results['dependencies']['msal'] = True
        verification_results['dependencies']['openai'] = True
        verification_results['dependencies']['httpx'] = True
    except ImportError as e:
        verification_results['dependencies']['missing'] = str(e)
    
    # Overall readiness check
    env_ready = all(verification_results['environment_variables'].values())
    services_ready = all(
        service.get('initialized', False) and service.get('mock_mode') == False
        for service in verification_results['services_ready'].values()
        if isinstance(service, dict)
    )
    deps_ready = 'missing' not in verification_results['dependencies']
    
    verification_results['overall_ready'] = env_ready and services_ready and deps_ready
    verification_results['status'] = 'ready' if verification_results['overall_ready'] else 'not_ready'
    
    return verification_results

if __name__ == "__main__":
    # When run directly, verify production readiness
    results = verify_production_readiness()
    print("Production Readiness Verification:")
    print("=" * 50)
    print(f"Overall Status: {results['status'].upper()}")
    print(f"Overall Ready: {results['overall_ready']}")
    print("\nEnvironment Variables:")
    for var, status in results['environment_variables'].items():
        print(f"  {var}: {'✓' if status else '✗'}")
    print("\nServices:")
    for service, status in results['services_ready'].items():
        if isinstance(status, dict):
            mock_status = "NO MOCK" if status.get('mock_mode') == False else "MOCK ENABLED" 
            print(f"  {service}: {'✓' if status.get('initialized') else '✗'} ({mock_status})")
        else:
            print(f"  Error: {status}")
    print("\nDependencies:")
    for dep, status in results['dependencies'].items():
        print(f"  {dep}: {'✓' if status else '✗'}")
