backend/requirements_integrations.txt
# Core dependencies for the Google Sheets and Microsoft Graph integration services
# These packages are required for the production-ready implementation

# Security and Cryptography
cryptography>=41.0.0
PyJWT>=2.8.0
bcrypt>=4.0.1

# HTTP and API clients
httpx>=0.25.0
requests>=2.31.0

# Google APIs
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
google-cloud-storage>=2.10.0

# Microsoft Graph
msal>=1.24.0
msgraph-core>=0.2.2

# Redis and Caching
redis>=5.0.0
hiredis>=2.2.3

# Async support
aioredis>=2.0.1
asyncio-throttle>=1.0.2

# Retry and resilience
tenacity>=8.2.0
backoff>=2.2.1

# Data validation and parsing
pydantic>=2.4.0
marshmallow>=3.20.0

# Flask extensions
Flask-JWT-Extended>=4.5.3
Flask-Limiter>=3.5.0
Flask-CORS>=4.0.0

# Database
SQLAlchemy>=2.0.0
alembic>=1.12.0

# Monitoring and logging
sentry-sdk>=1.32.0
structlog>=23.1.0

# Testing (development)
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-mock>=3.11.1
factory-boy>=3.3.0

# Development tools
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0
pre-commit>=3.4.0
